adding example BrowserRoute and Route
