+++ Curso-Redux-por-Bedu/blog/src/components/App.js
@@ -1,61 +1,16 @@
-import React, { Component } from 'react';
-import axios from 'axios';
+import { BrowserRouter, Route } from 'react-router-dom';
+import Menu from './Menu';
+import Usuarios from './Usuarios/index'
 
-class App extends Component {
-  constructor() {
-    super();
-    this.state = {
-      usuarios: []
-    }
-  }
+const Tareas = () => <div>Tareas</div>
 
-  async componentDidMount() {
-    const respuesta = await axios.get('https://jsonplaceholder.typicode.com/users');
-    //console.log('respuesta', respuesta.data)
-    this.setState({
-      usuarios: respuesta.data
-    });
-  }
+const App = () => (
 
-  ponerFilas = () => this.state.usuarios.map((usuario) => (
-    <tr key={usuario.id}>
-      <td>
-        {usuario.name}
-      </td>
-      <td>
-        {usuario.email}
-      </td>
-      <td>
-        {usuario.website}
-      </td>
-    </tr>
-  ));
-
-  render() {
-    //console.log(this.state.usuarios)
-    return (
-      <div className="margen">
-        <table className="tabla">
-          <thead>
-            <tr>
-              <th>
-                Nombre
-							</th>
-              <th>
-                Correo
-							</th>
-              <th>
-                Enlace
-							</th>
-            </tr>
-          </thead>
-          <tbody>
-            {this.ponerFilas()}
-          </tbody>
-        </table>
-      </div>
-    )
-  }
-};
+  <BrowserRouter>
+    <Menu />
+    <Route exact path='/' component={Usuarios} />
+    <Route exact path='/tareas' component={Tareas} />
+  </BrowserRouter>
+);
 
 export default App;
