+++ Curso-Redux-por-Bedu/blog/src/components/Menu.js
@@ -1,0 +1,14 @@
+import { Link } from 'react-router-dom';
+
+const Menu = (props) => (
+  <nav id='menu'>
+    <Link to='/'>
+      Usuarios
+		</Link>
+    <Link to='/tareas'>
+      Tareas
+		</Link>
+  </nav>
+);
+
+export default Menu;
