+++ Curso-Redux-por-Bedu/blog/src/index.css
@@ -25,3 +25,15 @@ body {
 	padding-top: 10px;
 	padding-bottom: 10px;
 }
+
+
+#menu {
+	background-color: #253A46;
+	padding: 20px;
+	font-size: 20px;
+}
+
+#menu a {
+	color: white;
+	padding-right: 50px;
+}
